backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
httpx
pytest
pyjwt
passlib[bcrypt]
bcrypt
python-jose
python-multipart
python-dotenv
alembic
psycopg2-binary